[BE] feat: 총 게임 결과 리턴 정보 추가

+++ backend/src/main/java/com/proj/withus/domain/GameResult.java
@@ -29,7 +29,7 @@ public class GameResult {
 
     private boolean isCorrect;
 
-    private int prediction;
+    private Long prediction;
 
-    private int answer;
+    private Long answer;
 }
